pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "lsy_drone_racing"
version = "0.0.1"
description = "LSY Autonomous Drone Racing Project Course"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
]

dependencies = [
    "fire >= 0.6.0",
    "munch >= 2.5.0",
    "numpy >= 1.24.1",
    "pandas >= 1.5.3",
    "pybullet >= 3.2.6",
    "PyYAML >= 6.0.1",
    "rospkg >= 1.5.1",
    "scipy >= 1.10.1",
]

[tool.setuptools]
packages = [
    "lsy_drone_racing", 
    "lsy_drone_racing.env_modifiers", 
    "controllers",
    "controllers.ppo"
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".git-rewrite",
    ".ipynb_checkpoints",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "D", "TCH", "ANN"]
ignore = ["ANN101", "ANN401"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true
suppress-none-returning = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
